♻️refactor: move image function to another file
