+++ package.json
@@ -16,8 +16,8 @@
     "react-hook-form": "^7.51.0"
   },
   "devDependencies": {
-    "@biomejs/biome": "^1.6.0",
-    "@happy-dom/global-registrator": "^13.7.8",
+    "@biomejs/biome": "^1.6.1",
+    "@happy-dom/global-registrator": "^13.8.2",
     "@heroicons/react": "^2.1.1",
     "@testing-library/react": "^14.2.1",
     "@types/bun": "^1.0.8",
